+++ app/models/log.rb
@@ -1,4 +1,6 @@
 class Log < ActiveRecord::Base
+  serialize :data, Hash
+
   belongs_to :loggable, polymorphic: true
 
   default_scope { order('created_at DESC') }