+++ db/migrate/20160303141417_add_data_to_logs.rb
@@ -1,0 +1,5 @@
+class AddDataToLogs < ActiveRecord::Migration
+  def change
+    add_column :logs, :data, :text, after: 'content'
+  end
+end