+++ changelog.md
@@ -4,6 +4,8 @@
 * Adds support for Spring in engine development
 * Expands Concerns::PaginationController with a method that handles both arrays
   and AR collections.
+* Allow page_number/per_page overrides through an options hash for 
+  Concerns::PaginationController#paginate
 * Add module for e-mail templates.
 * Add some styles for unordered lists in tables.
 * Add the Email model which holds all the sent/unsent e-mails.
@@ -19,6 +21,7 @@
 * Tweak css for the search filter panels.
 * Add QueuedTask.queue_unless_already_queued which speaks for itself.
 * Add .icons class for the <td> which contains icons.
+* Add 'data' field to Log which contains a an optional hash.
 
 
 0.0.12 - 2016-01-22